requirements.txt
telethon>=1.24.0
pysocks>=1.7.1 